update of services empresa parking
actualización de servicios relacionados a empresa y parqueaderos

+++ src/main/java/com/parkinggo/repository/IDetalleFacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.parkinggo.model.DetalleFactura;
 
-public interface IDetalleFacturaRepository extends JpaRepository<DetalleFactura, Integer>{
+public interface IDetalleFacturaRepository extends JpaRepository<DetalleFactura, Long>{
 
 }
