+++ src/main/java/com/parkinggo/repository/IParqueaderoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.parkinggo.model.Parqueadero;
 
-public interface IParqueaderoRepository extends JpaRepository<Parqueadero, Integer>{
+public interface IParqueaderoRepository extends JpaRepository<Parqueadero, Long>{
 
 }
