+++ src/main/java/com/parkinggo/repository/IPuertaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.parkinggo.model.Puerta;
 
-public interface IPuertaRepository extends JpaRepository<Puerta, Integer>{
+public interface IPuertaRepository extends JpaRepository<Puerta, Long>{
 
 }
